move FloodgateInfoHolder to dump package in connector
